issue #113 refactor: MomentwoApi/Auth/ApiModule/AuthModule 을 :core:network 로 이동
HTTP 인프라(MomentwoApi 엔드포인트 상수, AuthInterceptor/AuthAuthenticator,
ApiModule, AuthModule)를 :core:network 로 이동한다. BASE_URL 은 network 의
BuildConfig 로 옮기고 app 의 미사용 BASE_URL 설정을 제거했다. AuthInterceptor 가
PreferenceRepository 를 참조하므로 network 는 :core:datastore 에 의존한다.
DTO/Service/DataSource 는 매핑과 한 덩어리라 :core:data PR 에서 함께 이동한다.
