+++ app/src/main/java/cord/eoeo/momentwo/data/login/remote/LoginService.kt
@@ -1,6 +1,6 @@
 package cord.eoeo.momentwo.data.login.remote
 
-import cord.eoeo.momentwo.data.MomentwoApi
+import cord.eoeo.momentwo.core.network.MomentwoApi
 import cord.eoeo.momentwo.data.model.LoginRequest
 import cord.eoeo.momentwo.data.model.UserProfile
 import retrofit2.Response
