+++ app/src/main/java/cord/eoeo/momentwo/data/photo/remote/PhotoService.kt
@@ -1,6 +1,6 @@
 package cord.eoeo.momentwo.data.photo.remote
 
-import cord.eoeo.momentwo.data.MomentwoApi
+import cord.eoeo.momentwo.core.network.MomentwoApi
 import cord.eoeo.momentwo.data.model.LikedPhotoList
 import cord.eoeo.momentwo.data.model.PhotoPage
 import cord.eoeo.momentwo.data.model.PresignedRequest
