+++ app/src/main/java/cord/eoeo/momentwo/data/profile/remote/ProfileService.kt
@@ -1,6 +1,6 @@
 package cord.eoeo.momentwo.data.profile.remote
 
-import cord.eoeo.momentwo.data.MomentwoApi
+import cord.eoeo.momentwo.core.network.MomentwoApi
 import cord.eoeo.momentwo.data.model.UserProfile
 import retrofit2.http.GET
 import retrofit2.http.Query
