+++ core/network/src/main/java/cord/eoeo/momentwo/core/network/AuthAuthenticator.kt
@@ -1,4 +1,4 @@
-package cord.eoeo.momentwo.data.authentication
+package cord.eoeo.momentwo.core.network
 
 import android.util.Log
 import cord.eoeo.momentwo.core.datastore.PreferenceRepository
